Se mejoro sistema de seguimiento

+++ coe/app/functions.py
@@ -15,25 +15,31 @@ def obtener_dni(data):
 #Funcionalidades
 def activar_tracking(individuo):
     try:
-        AppNotificacion.objects.filter(appdata=individuo.appdata).delete()
-        notif = AppNotificacion(appdata=individuo.appdata)
-        notif.titulo = 'Iniciar Proceso de supervisión Digital'
-        notif.mensaje = 'Por Favor presione esta notificacion para iniciarlo.'
-        notif.accion = 'BT'
-        notif.save()#Al grabar el local, se envia automaticamente por firebase (signals)
-    except:
+        if individuo.appdata:
+            print("Iniciar Tracking: " + str(individuo))
+            AppNotificacion.objects.filter(appdata=individuo.appdata).delete()
+            notif = AppNotificacion(appdata=individuo.appdata)
+            notif.titulo = 'Iniciar Proceso de supervisión Digital'
+            notif.mensaje = 'Por Favor presione esta notificacion para iniciarlo.'
+            notif.accion = 'BT'
+            notif.save()#Al grabar el local, se envia automaticamente por firebase (signals)
+            print("Guardado")
+    except Exception as e:
         logger.info("\nFallo Activar Tracking: "+str(individuo))
+        logger.info(e)
         logger.info(traceback.format_exc())
 
 def desactivar_tracking(individuo):
     try:
-        AppNotificacion.objects.filter(appdata=individuo.appdata).delete()
-        notif = AppNotificacion(appdata=individuo.appdata)
-        notif.titulo = 'Finalizar Proceso de supervisión Digital'
-        notif.mensaje = 'Por Favor presione esta notificacion para terminarlo.'
-        notif.accion = 'ST'
-        notif.save()#Al grabar el local, se envia automaticamente por firebase (signals)
-    except:
+        if individuo.appdata:
+            AppNotificacion.objects.filter(appdata=individuo.appdata).delete()
+            notif = AppNotificacion(appdata=individuo.appdata)
+            notif.titulo = 'Finalizar Proceso de supervisión Digital'
+            notif.mensaje = 'Por Favor presione esta notificacion para terminarlo.'
+            notif.accion = 'ST'
+            notif.save()#Al grabar el local, se envia automaticamente por firebase (signals)
+    except Exception as e:
         logger.info("\nFallo Desactivar Tracking: "+str(individuo))
+        logger.info(e)
         logger.info(traceback.format_exc())
